requirements.txt
Django<4.0
djangorestframework<4.0
Pillow<8